+++ src/components/Field/index.ts
@@ -1,3 +1,4 @@
-// Field from copy-ui @ 2025-11-08.
+// Field from copy-ui @ 2025-12-17.
 
 export { Field } from "./field";
+export { useOptionalFieldContext } from "./field-context";